gateway: Move config module into lib.rs
